Add maxSubscriber limit to Completable/Single Processors
Motivation:
`Processors.new[Single|Completable]Processor` create a `Processor`s
that permit an unbounded number of `Subscriber`s. This may result
in memory pressure in some scenarios where the number of `Subscriber`s
is high.

Modifications:
- `ClosableConcurrentStack` to support `maxSize` to bound
  memory for its use cases.
- Add new methods for `Processors.new[Single|Completable]Processor`
  that allow configuring the maximum number of `Subscriber`s.

Result:
Completable/Single Processors and underlying data structure support
limiting the maximum amount of `Subscriber`s.

+++ servicetalk-concurrent-api/src/main/java/io/servicetalk/concurrent/api/IterableMergeCompletable.java
@@ -37,7 +37,7 @@ final class IterableMergeCompletable extends AbstractMergeCompletableOperator<Co
 
     @Override
     void doMerge(final CompletableDynamicCountSubscriber subscriber) {
-        long count = 1;
+        int count = 1;
         for (Completable itr : others) {
             ++count;
             itr.subscribeInternal(subscriber);
